#620: Demuck config so we can decouple the things
