+++ lib/bootstrap.js
@@ -7,9 +7,17 @@
  *
  * // Get the bootstrap function
  * var bootstrap = require('./../lib/bootstrap.js');
+ * var options = {
+ *   logLevelConsole: LOGLEVELCONSOLE,
+ *   userConfRoot: USERCONFROOT,
+ *   envPrefix: ENVPREFIX,
+ *   configSources: configSources,
+ *   pluginDirs: [USERCONFROOT],
+ *   mode: 'cli'
+ * };
  *
- * // Initialize Lando in CLI mode
- * bootstrap({mode: 'cli'})
+ * // Initialize Lando with some options
+ * bootstrap(options)
  *
  * // Initialize some other thing
  * .tap(function(lando) {
@@ -21,10 +29,6 @@
 
 // Modules
 var _ = require('lodash');
-var config = require('./config');
-var SILLY_TEXT = 'It\'s not particularly silly, is it? I mean, the right leg' +
-  ' isn\'t silly at all and the left leg merely does a forward aerial half' +
-  ' turn every alternate step.';
 
 /**
  * The main bootstrap function.
@@ -35,40 +39,62 @@ var SILLY_TEXT = 'It\'s not particularly silly, is it? I mean, the right leg' +
  *   2. Emit bootstrap events
  *   3. Initialize plugins
  *
+ * The intiialization config below is not required but recommended. You can
+ * pass in any additional properties to override subsequently set/default values.
+ *
+ * Check out bin/lando.js for an example
+ *
  * @since 3.0.0
  * @name bootstrap
  * @static
  * @fires pre-bootstrap
  * @fires post-bootstrap
- * @param {Object} opts - Options to tweak the bootstrap
- * @param {String} opts.mode - The mode to run the bootstrap with
+ * @param {Object} [options] - Options to initialize the bootstrap
  * @returns {Object} An initialized Lando object
  * @example
  *
  * // Get the bootstrap function
  * var bootstrap = require('./../lib/bootstrap.js');
  *
- * // Initialize Lando in CLI mode
- * bootstrap({mode: 'cli'})
+ * // Initialize Lando with some start up config
+ * bootstrap({
+ *   logLevelConsole: LOGLEVELCONSOLE,
+ *   userConfRoot: USERCONFROOT,
+ *   envPrefix: ENVPREFIX,
+ *   configSources: configSources,
+ *   pluginDirs: [USERCONFROOT],
+ *   mode: 'cli'
+ * })
  *
  * // Initialize CLI
  * .tap(function(lando) {
  *   return lando.cli.init(lando);
  * })
  */
-module.exports = _.once(function(opts) {
+module.exports = _.once(function(options) {
 
-  // @TODO: Mixing and validate config with new config mods
+  // Get our config helpers
+  var helpers = require('./config');
 
-  // Merge in our opts to the configs
-  config = _.merge(config, opts);
+  // Start building the config
+  var config = helpers.merge(helpers.defaults(), options);
+
+  // If we have configSources let's merge those in as well
+  if (!_.isEmpty(config.configSources)) {
+    config = helpers.merge(config, helpers.loadFiles(config.configSources));
+  }
+
+  // If we have an envPrefix set then lets merge that in as well
+  if (_.has(config, 'envPrefix')) {
+    config = helpers.merge(config, helpers.loadEnvs(config.envPrefix));
+  }
 
   // Summon lando
   var lando = require('./lando')(config);
 
   // Log that we've begun
   lando.log.info('Bootstraping...');
-  lando.log.silly(SILLY_TEXT);
+  lando.log.silly('It\'s not particularly silly, is it?');
 
   /**
    * Event that allows other things to augment the lando global config.
