+++ lib/cli.js
@@ -1,21 +1,198 @@
 /**
- * Contains some CLI artwork and formatting helpers
+ * Contains some CLI artwork and helpers
  *
  * @since 3.0.0
  * @module cli
  * @example
  *
+ * // Gets all the global options that have been specified.
+ * var largv = lando.tasks.largv;
+ *
  * // Get the start header
  * return lando.cli.startHeader();
+ *
+ * // Add a task so it shows up as a command in the CLI
+ * yargs.command(lando.tasks.parseToYargs(task));
+ *
  */
 
 'use strict';
 
 // Modules
-var _ = require('./node')._;
+var _ = require('lodash')._;
 var chalk = require('./node').chalk;
+var inquirer = require('inquirer');
 var os = require('os');
+var parse = require('yargs-parser');
 var Table = require('cli-table');
+
+/**
+ * Returns the lando options
+ *
+ * This means all the options passed in before the `--` flag.
+ *
+ * @since 3.0.0
+ * @example
+ *
+ * // Gets all the global options that have been specified.
+ * var argv = lando.tasks.argv;
+ */
+exports.argv = function() {
+  return require('yargs').argv;
+};
+
+/**
+ * Helper function to parse global opts
+ *
+ * @since 3.0.0
+ * @see https://github.com/lando/lando/issues/351
+ * @example
+ *
+ * // Gets all the tasks that have been loaded
+ * var largv = lando.tasks.parseGlobals();
+ */
+exports.parseGlobals = function() {
+
+  // Get our raw
+  var rawArgs = process.argv.slice(2);
+  var globz = parse(rawArgs);
+
+  // The first parsing
+  // Check if we have _. and then filter by -
+  if (!_.isEmpty(globz._)) {
+    globz = _.filter(globz._, function(value) {
+      return _.startsWith(value, '-');
+    });
+  }
+  else {
+    globz = [];
+  }
+
+  // Return the globalz
+  return parse(globz, {
+    alias: {'verbose': ['v'], 'help': ['h']},
+    count: ['v']
+  });
+
+};
+
+/**
+ * A singleton object that contains the Lando global options.
+ *
+ * This means all the options passed in after the `--` flag.
+ *
+ * @since 3.0.0
+ * @example
+ *
+ * // Gets all the global options that have been specified.
+ * var largv = lando.tasks.largv;
+ */
+exports.largv = this.parseGlobals();
+
+/**
+ * Parses a lando task object into something that can be used by the [yargs](http://yargs.js.org/docs/) CLI.
+ *
+ * A lando task object is an abstraction on top of yargs that also contains some
+ * metadata about how to interactively ask questions on both the CLI and GUI. While this
+ * method is useful, any task added to Lando via `lando.tasks.add` will automatically
+ * be parsed with this method.
+ *
+ * The interactivity metadata is light wrapper around [inquirer](https://github.com/sboudrias/Inquirer.js)
+ *
+ * @since 3.0.0
+ * @see [yargs docs](http://yargs.js.org/docs/)
+ * @see [inquirer docs](https://github.com/sboudrias/Inquirer.js)
+ * @param {Object} task - A Lando task object (@see add for definition)
+ * @returns {Object} A yargs command object
+ * @example
+ *
+ * // Add that task to the CLI
+ * yargs.command(lando.tasks.parseToYargs(task));
+ */
+exports.parseToYargs = function(task) {
+
+  // Start command and option collectors
+  var command = {};
+  var inquiry = [];
+
+  // Grab the basics
+  command.command = task.command;
+  command.describe = task.describe;
+
+  // Add our options to the builder
+  command.builder = task.options || {};
+
+  // Translate options to inquiry
+  _.forEach(command.builder, function(option, name) {
+    if (option && option.interactive) {
+      var question = option.interactive;
+      question.name = question.name || name;
+      question.weight = question.weight || 0;
+      inquiry.push(question);
+    }
+  });
+
+  // Build a handler that incorporates interactive options
+  command.handler = function(argv) {
+
+    // Get lando
+    var lando = require('./lando')(require('./config'));
+    var cmd = argv._[0];
+    var answersEvent = ['task', cmd, 'answers'].join('-');
+    var runEvent = ['task', cmd, 'run'].join('-');
+
+    /**
+     * Event that allows altering of argv or inquirer before interactive prompts
+     * are run
+     *
+     * @since 3.0.0
+     * @event module:tasks.event:task-CMD-answers
+     * @property {Object} answers argv and inquirer questions
+     */
+    return lando.events.emit(answersEvent, {argv: argv, inquirer: inquiry})
+
+    // Do the filtering
+    .then(function() {
+
+      // Attempt to filter out questions that have already been answered
+      var questions = _.filter(inquiry, function(option) {
+        return (_.isNil(argv[option.name]) || argv[option.name] === false);
+      });
+
+      // Prompt the use for feedback if needed and sort by weight
+      return inquirer.prompt(_.sortBy(questions, 'weight'));
+
+    })
+
+    // Merge data from our interfaces and provide to the task runner
+    .then(function(answers) {
+
+      // Merge it all together
+      var truth = _.merge(argv, answers);
+
+      /**
+       * Event that allows final altering of answers
+       *
+       * @since 3.0.0
+       * @event module:tasks.event:task-CMD-run
+       * @property {Object} answers object
+       */
+      return lando.events.emit(runEvent, truth)
+
+      // Run the task
+      .then(function() {
+        return task.run(truth);
+      });
+
+    });
+
+  };
+
+  // Return the completed object
+  return command;
+
+};
+
 
 /**
  * Returns a cheeky header that can be used after an app is started.
