+++ lib/config.js
@@ -1,57 +1,98 @@
 /**
- * Generates and returns the global config object
+ * Helpers to build the initialization configuration.
+ *
+ * These are meant to be used to help your entrypoint script build a config
+ * object
  *
  * @since 3.0.0
  * @module config
  * @example
  *
- * // Get the config
- * var config = lando.config;
+ * // Get the config helpers
+ * var config = require('./config');
  */
 
 'use strict';
 
 // Modules
 var _ = require('lodash');
-var env = require('./env');
 var fs = require('fs-extra');
 var os = require('os');
 var path = require('path');
-//var user = require('./user');
 var yaml = require('js-yaml');
+var _config = this;
 
-// Set our deepest of deep core config
-//var APP_ROOT_DIRNAME = process.env.LANDO_CORE_APP_ROOT_DIRNAME || 'Lando';
-var CONFIG_FILENAME = process.env.LANDO_CORE_CONFIG_FILENAME || 'config.yml';
-var ENV_PREFIX = process.env.LANDO_CORE_ENV_PREFIX || 'LANDO_';
-//var LANDOFILE_NAME = process.env.LANDO_CORE_LANDOFILE_NAME || '.lando.yml';
+/**
+ * Document
+ */
+var getSysConfRoot = function() {
+
+  // Win path
+  var win = process.env.LANDO_INSTALL_PATH || 'C:\\Program Files\\Lando';
+
+  // Return sysConfRoot based on path
+  switch (process.platform) {
+    case 'win32': return win;
+    case 'darwin': return '/Applications/Lando.app/Contents/MacOS';
+    case 'linux': return '/usr/share/lando';
+  }
+
+};
+
+/**
+ * Config merging utility
+ */
+exports.merge = function(old, fresh) {
+  return _.mergeWith(old, fresh, function(s, f) {
+    if (_.isArray(s)) {
+      return s.concat(f);
+    }
+  });
+};
 
 /*
  * Define default config
  */
-var getDefaultConfig = function() {
+exports.defaults = function() {
 
   // Grab version things
-  var packageJson = require(path.join(env.srcRoot, 'package.json'));
+  var configFilename = 'config.yml';
+  var packageJson = require(path.join(__dirname, '..', 'package.json'));
+  var srcRoot = path.resolve(__dirname, '..');
+
+  // Load up our initial env
+  var env = process.env || {};
+
+  // Set Path environmental variable if we are on windows so we get access
+  // to things like ssh.exe
+  // @TODO: is all of this still needed?
+  if (process.platform === 'win32') {
+    var appData = process.env.LOCALAPPDATA;
+    var programFiles = process.env.ProgramFiles;
+    var programFiles2 = process.env.ProgramW6432;
+    var gitBin1 = path.join(appData, 'Programs', 'Git', 'usr', 'bin');
+    var gitBin2 = path.join(programFiles, 'Git', 'usr', 'bin');
+    var gitBin3 = path.join(programFiles2, 'Git', 'usr', 'bin');
+
+    // Only add the gitbin to the path if the path doesn't start with
+    // it. We want to make sure gitBin is first so other things like
+    // putty don't F with it.
+    //
+    // See https://github.com/kalabox/kalabox/issues/342
+    _.forEach([gitBin1, gitBin2, gitBin3], function(gBin) {
+      if (fs.existsSync(gBin) && !_.startsWith(env.Path, gBin)) {
+        env.Path = [gBin, env.Path].join(';');
+      }
+    });
+  }
 
   var config = {
-    //appConfigFilename: LANDOFILE_NAME,
-    //appsRoot: path.join(env.home, APP_ROOT_DIRNAME),
-    //appRegistry: path.join(env.userConfRoot, 'appRegistry.json'),
-    //cache: true,
-    //composeBin: env.composeBin,
-    //composeVersion: '3.2',
-    configFilename: CONFIG_FILENAME,
-    //containerGlobalEnv: {},
-    //dockerBin: env.dockerBin,
-    //dockerBinDir: env.dockerBinDir,
-    //engineId: user.getEngineUserId(),
-    //engineGid: user.getEngineUserGid(),
-    env: env.env,
-    home: env.home,
+    configFilename: configFilename,
+    configSources: [path.join(srcRoot, configFilename)],
+    env: env,
+    home: os.homedir(),
     logLevel: 'debug',
-    logLevelConsole: 'info',
-    //loadPassphraseProtectedKeys: false,
+    logLevelConsole: 'warn',
     node: process.version,
     os: {
       type: os.type(),
@@ -59,18 +100,10 @@ var getDefaultConfig = function() {
       release: os.release(),
       arch: os.arch()
     },
-    pluginDirs: [
-      env.srcRoot,
-      env.sysConfRoot,
-      env.userConfRoot
-    ],
-    stats: [{
-      report: true,
-      url: 'http://metrics.devwithlando.io'
-    }],
-    srcRoot: env.srcRoot,
-    sysConfRoot: env.sysConfRoot,
-    userConfRoot: env.userConfRoot,
+    pluginDirs: [srcRoot],
+    srcRoot: srcRoot,
+    sysConfRoot: getSysConfRoot(),
+    userConfRoot: path.join(os.homedir(), '.lando'),
     version: packageJson.version
   };
 
@@ -82,33 +115,36 @@ var getDefaultConfig = function() {
 /*
  * Merge in config file if it exists
  */
-var mergeConfigFile = function(file) {
+exports.loadFiles = function(files) {
 
-  // Merge in file if it exists
-  if (fs.existsSync(file)) {
-    return yaml.safeLoad(fs.readFileSync(file));
-  }
+  // Start collecting
+  var conf = {};
 
-  // Return emptiness
-  return {};
+  // Merge all the files on top of one another, last file is last merged and
+  // has precedent
+  _.forEach(files, function(file) {
+    if (fs.existsSync(file)) {
+      conf = _config.merge(conf, yaml.safeLoad(fs.readFileSync(file)));
+    }
+  });
+
+  // Return
+  return conf;
 
 };
 
 /*
- * Merge in env vars if they exists
+ * Grab envvars and map to config
  */
-var mergeEnvVars = function() {
+exports.loadEnvs = function(prefix) {
 
   // Start an object of our env
   var envConfig = {};
 
   // Build object of lando_ envvars
   _.forEach(process.env, function(value, key) {
-    if (_.includes(key, ENV_PREFIX)) {
-
-      // Add to our config object
-      envConfig[_.camelCase(_.trimStart(key, ENV_PREFIX))] = value;
-
+    if (_.includes(key, prefix)) {
+      envConfig[_.camelCase(_.trimStart(key, prefix))] = value;
     }
   });
 
@@ -116,33 +152,3 @@ var mergeEnvVars = function() {
   return envConfig;
 
 };
-
-/*
- * Get config
- */
-var getConfig = _.memoize(function() {
-
-  // Start with the default config
-  var config = getDefaultConfig();
-
-  // Check for other config soruces
-  var configSources = [
-    path.join(config.srcRoot, CONFIG_FILENAME),
-    path.join(config.sysConfRoot, CONFIG_FILENAME),
-    path.join(config.userConfRoot, CONFIG_FILENAME)
-  ];
-
-  // Check to see if we have more config files and merge those in
-  _.forEach(configSources, function(file) {
-    config = _.merge(config, mergeConfigFile(file));
-  });
-
-  // Check for ENV values and merge those in as well
-  return _.merge(config, mergeEnvVars());
-
-});
-
-/*
- * Export func to get conf
- */
-module.exports = getConfig();
