+++ lib/lando.js
@@ -46,7 +46,6 @@
 var _ = require('lodash');
 var AsyncEvents = require('./events');
 var Log = require('./logger');
-var path = require('path');
 
 /*
  * Return lando
@@ -56,11 +55,7 @@ module.exports = _.memoize(function(config) {
   // Instantiate a logger so we can set it here and inject it into our
   // other modules as needed
   /// @TODO: some better validation here?
-  var log = new Log({
-    logLevel: config.logLevel,
-    logLevelConsole: config.logLevelConsole,
-    logDir: path.join(config.userConfRoot, 'logs')
-  });
+  var log = new Log(config);
 
   // Create the lando object
   var lando = {
@@ -85,7 +80,7 @@ module.exports = _.memoize(function(config) {
      * @memberof lando
      * @see {@link cli.md}
      */
-    //cli: require('./cli'),
+    cli: require('./cli'),
 
     /**
      * The global config object
