+++ lib/logger.js
@@ -116,9 +116,24 @@ var winston = require('winston');
  */
 function Log(opts) {
 
+  // Define the loglevels
+  var logLevels = {
+    '0': 'error',
+    '1': 'warn',
+    '2': 'info',
+    '3': 'verbose',
+    '4': 'debug',
+    '5': 'silly'
+  };
+
   // Some overrideable default options
   var logLevelConsole = opts.logLevelConsole || 'warn';
   var logLevel = opts.logLevel || 'debug';
+
+  // If loglevelconsole is numeric lets map it!
+  if (_.isInteger(logLevelConsole)) {
+    logLevelConsole = logLevels[logLevelConsole] || 'warn';
+  }
 
   // The default console transport
   var transports = [
