+++ plugins/lando-core/lib/registry.js
